Refactored some namespaces and changed the CommandExtenderServiceProvider.php to loop through the extended commands.
