+++ src/CommandExtenderServiceProvider.php
@@ -3,12 +3,16 @@
 namespace Wotta\CommandExtender;
 
 use Illuminate\Support\ServiceProvider;
-use Wotta\CommandExtender\Console\ExtendedRouteListCommand;
+use Wotta\CommandExtender\Console\Extended\RouteListCommand;
 
 class CommandExtenderServiceProvider extends ServiceProvider
 {
     protected $commands = [
-        ExtendedRouteListCommand::class,
+        RouteListCommand::class,
+    ];
+
+    protected $extendedCommands = [
+        'command.route.list' => RouteListCommand::class,
     ];
 
     public function boot(): void
@@ -34,8 +38,11 @@ class CommandExtenderServiceProvider extends ServiceProvider
             return new CommandExtender;
         });
 
-        $this->app->extend('command.route.list', function () {
-            return $this->app->make(ExtendedRouteListCommand::class);
-        });
+        foreach ($this->extendedCommands as $abstract => $extendedCommand) {
+            $this->app->extend($abstract, function () use ($extendedCommand) {
+                return $this->app->make($extendedCommand);
+            });
+        }
+
     }
 }
